+++ CHANGELOG.md
@@ -1,5 +1,27 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r226-release...07499d9388070c25ad756c32d2b5dd94c3e226df)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r227-release...9ac8f6f2e810167bc39007b18805997325dacc37)
 
-- Fix QuickAssist font shadows
-- Fix indicator updates
-- Update zhTW
+Vollmer (2):
+
+- Add ability to hide non selected indicators in preview (#125)
+- Update bleedList (#119)
+
+enderneko (35):
+
+- Implement Icons Spacing
+- Implement "Block" indicator type
+- Implement "Sort by Role per Group"
+- Improve usability of Custom Nicknames
+- Add "Enabled" for ReadyCheckIcon
+- Add "Solo" for LayoutAutoSwitch
+- Add duration text for Bar/Rect
+- Implemented "Use Elapsed Time"
+- Update RaidDebuffs DB structure
+- Update Cata Raid Debuffs
+- Update TargetedSpells
+- Raise the frame level of Dispel Highlight
+- Optimize indicator OnUpdate
+- Fix Spotlight menu frame level
+- Fix dispel checker for cata
+- Fix Click-Castings
+- Fix "Command" keybinding
+- Update Supporters